CID-338 moved helper to Lib dir
